Renamed value renderer
