+++ superset-frontend/src/cccs-viz/plugins/plugin-chart-ag-grid/src/rendererMap.ts
@@ -3,7 +3,7 @@ import DomainValueRenderer from './renderers/DomainValueRenderer';
 import Ipv4ValueRenderer from './renderers/Ipv4ValueRenderer';
 import Ipv6ValueRenderer from './renderers/Ipv6ValueRenderer';
 import JsonValueRenderer from './renderers/JsonValueRenderer';
-import TimestampValueRenderer from './renderers/TimestampWithoutTimezoneValueRenderer';
+import TimestampValueRenderer from './renderers/TimestampValueRenderer';
 
 // Key is column advanced type, value is renderer
 const rendererMap = {
